% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/containerwidget.R
\name{renderContainerwidget}
\alias{renderContainerwidget}
\title{Widget render function for use in Shiny}
\usage{
renderContainerwidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

